Move the geo APIs to account
