+++ README.md
@@ -21,9 +21,9 @@ npm install @dolbyio/dolbyio-rest-apis-client
 ## Create a publish token
 
 ```ts
-import { streaming } from '@dolbyio/dolbyio-rest-apis-client'
+import { streaming } from '@dolbyio/dolbyio-rest-apis-client';
 
-const API_KEY = process.env.DOLBYIO_API_SECRET;;
+const API_KEY = process.env.DOLBYIO_API_SECRET;
 
 const publishToken = await streaming.publishToken.create(API_KEY, {
     label: 'My token',
