+++ src/streaming/index.ts
@@ -12,7 +12,6 @@ export * as cluster from './cluster';
  */
 export * as transcoders from './transcoders';
 export * as director from './director';
-export * as geo from './geo';
 export * as publishToken from './publishToken';
 export * as recordFiles from './recordFiles';
 export * as stream from './stream';
