+++ src/streaming/types/account.ts
@@ -14,3 +14,13 @@ export interface AccountGeoCascade {
      */
     clusters?: string[];
 }
+
+/**
+ * Represents the geo restrictions rules.
+ */
+export interface GeoRestrictions {
+    /** List of allowed countries. */
+    allowedCountries: string[];
+    /** List of restricted countries. */
+    deniedCountries: string[];
+}
