+++ src/streaming/types/index.ts
@@ -1,7 +1,6 @@
 export * from './account';
 export * from './cluster';
 export * from './director';
-export * from './geo';
 export * from './publishToken';
 export * from './recordFiles';
 export * from './stream';
